# remove callable args alloc