@geseq 2020-03-15
Yes looks good. It reminds me of a bunch of optimizations I pushed to the tengo2 branch. I’ll look into those as well